tuck userProfile into myDashboard
