+++ api/serializers.py
@@ -47,6 +47,7 @@ class AuthenticatedUserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = [
+            "id",
             "first_name",
             "last_name",
             "username",
